Fix compiler warnings, turn on more optimisation settings in release config

+++ Asteroids/src/Config.h
@@ -2,7 +2,7 @@
 
 namespace config
 {
-    static const float SCR_WIDTH(1024);
-    static const float SCR_HEIGHT(768);
+    static const int SCR_WIDTH(1024);
+    static const int SCR_HEIGHT(768);
     static const double SPF(1.0 / 60.0);
 };
