+++ Asteroids/src/renderer/Renderer.cpp
@@ -11,7 +11,7 @@ Renderer::Renderer(GLFWwindow* window, std::string vertexPath, std::string fragm
     // Initialise camera
     view = glm::translate(view, glm::vec3(0.0f, 0.0f, -32.0f));
 
-    projection = glm::ortho(0.0f, config::SCR_WIDTH, 0.0f, config::SCR_HEIGHT, 0.1f, 100.0f);
+    projection = glm::ortho(0.0f, static_cast<float>(config::SCR_WIDTH), 0.0f, static_cast<float>(config::SCR_HEIGHT), 0.1f, 100.0f);
     //projection = glm::ortho(-config::SCR_WIDTH, config::SCR_WIDTH * 2, -config::SCR_HEIGHT, config::SCR_HEIGHT * 2, 0.1f, 100.0f);
 
     // Initialise shader
